add security and hateoas

+++ src/Controller/UserController.php
@@ -35,12 +35,36 @@ class UserController extends AbstractController
      *     @OA\Items(ref=@Model(type=User::class, groups={"getUserDetail"}))
      *    )
      * )
+     * @OA\Response(
+     *       response=401,
+     *       description="Non authentifié",
+     *   )
+     * @OA\Response(
+     *     response=403,
+     *     description="Accès refusé",
+     *     )
+     * @OA\Response(
+     *       response=404,
+     *       description="Resource non trouvée",
+     *   )
+     * @OA\Response(
+     *      response=500,
+     *      description="Erreur serveur",
+     *      )
+     *
      * @OA\Tag(name="Users")
      *
      */
     #[Route('api/users/{id}', name: 'app_user_detail', methods: ['GET'])]
     public function getUserDetail(User $user, SerializerInterface $serializer): JsonResponse
     {
+        if ($user->getCustomer() !== $this->getUser()) {
+            return new JsonResponse(
+                null,
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
         $context = SerializationContext::create()->setGroups(['getUserDetail']);
         $jsonUser = $serializer->serialize($user, 'json', $context);
 
@@ -57,12 +81,35 @@ class UserController extends AbstractController
      *     response=204,
      *     description="Supprime un utilisateur",
      * )
+     * @OA\Response(
+     *        response=401,
+     *        description="Non authentifié",
+     *    )
+     * @OA\Response(
+     *      response=403,
+     *      description="Accès refusé",
+     *      )
+     * @OA\Response(
+     *        response=404,
+     *        description="Resource non trouvée",
+     *    )
+     * @OA\Response(
+     *      response=500,
+     *      description="Erreur serveur",
+     *      )
      * @OA\Tag(name="Users")
      *
      */
     #[Route('api/users/{id}', name: 'user_delete', methods: ['DELETE'])]
     public function deleteUser(User $user, EntityManagerInterface $em, TagAwareCacheInterface $cache): JsonResponse
     {
+        if ($user->getCustomer() !== $this->getUser()) {
+            return new JsonResponse(
+                null,
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
         $cache->invalidateTags(['usersCache' . $user->getCustomer()->getId()]);
         $em->remove($user);
         $em->flush();
@@ -76,58 +123,55 @@ class UserController extends AbstractController
 
     /**
      * @throws InvalidArgumentException
+     *
      * @OA\Response(
      *     response=204,
      *     description="Met à jour un utilisateur",
      * )
+     * @OA\Response(
+     *        response=401,
+     *        description="Non authentifié",
+     *    )
+     * @OA\Response(
+     *      response=403,
+     *      description="Accès refusé",
+     *      )
+     * @OA\Response(
+     *        response=404,
+     *        description="Resource non trouvée",
+     *    )
+     * @OA\Response(
+     *      response=500,
+     *      description="Erreur serveur",
+     *      )
      * @OA\Tag(name="Users")
      * @OA\RequestBody(
      *      @OA\JsonContent(
      *           example={
      *              "email": "email@test.fr",
-     *              "password": "password",
-     *              "firstName": "firstName",
-     *              "lastName": "lastName",
-     *              "customerId": 1
+     *              "firstname": "firstname",
+     *              "lastname": "lastname",
      *           },
      *      )
      * )
-     * @OA\Parameter(
-     *     name="email",
-     *     in="query",
-     *     required=false,
-     *     description="Email de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="password",
-     *     required=false,
-     *     in="query",
-     *     description="Mot de passe de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="firstName",
-     *     in="query",
-     *     description="Prénom de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *    name="lastName",
-     *     in="query",
-     *     description="Nom de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="customerId",
-     *     in="query",
-     *     description="Id du customer",
-     *     @OA\Schema(type="integer")
-     * )
      */
     #[Route('api/users/{id}', name: 'user_update', methods: ['PUT'])]
-    public function updateUser(User $currentUser, SerializerInterface $serializer, EntityManagerInterface $em, Request $request, CustomerRepository $customerRepository, ValidatorInterface $validator, TagAwareCacheInterface $cache, UserPasswordHasherInterface $passwordHasher): JsonResponse
-    {
+    public function updateUser(
+        User                        $currentUser,
+        SerializerInterface         $serializer,
+        EntityManagerInterface      $em,
+        Request                     $request,
+        CustomerRepository          $customerRepository,
+        ValidatorInterface          $validator,
+        TagAwareCacheInterface      $cache,
+        UserPasswordHasherInterface $passwordHasher
+    ): JsonResponse {
+        if ($currentUser->getCustomer() !== $this->getUser()) {
+            return new JsonResponse(
+                null,
+                Response::HTTP_FORBIDDEN
+            );
+        }
 
         $updatedUser = $serializer->deserialize($request->getContent(), User::class, 'json');
         //on met a jour uniquement les champs modifiés
@@ -136,25 +180,14 @@ class UserController extends AbstractController
             $currentUser->setEmail($updatedUser->getEmail());
         }
 
-        if ($updatedUser->getRoles() !== null) {
-            $currentUser->setRoles($updatedUser->getRoles());
+        if ($updatedUser->getFirstname() !== null) {
+            $currentUser->setFirstname($updatedUser->getFirstname());
         }
 
-        if ($updatedUser->getCustomer() !== null) {
-            $currentUser->setCustomer($updatedUser->getCustomer());
+        if ($updatedUser->getLastname() !== null) {
+            $currentUser->setLastname($updatedUser->getLastname());
         }
 
-        if ($updatedUser->getFirstName() !== null) {
-            $currentUser->setFirstName($updatedUser->getFirstName());
-        }
-
-        if ($updatedUser->getLastName() !== null) {
-            $currentUser->setLastName($updatedUser->getLastName());
-        }
-
-        if ($updatedUser->getPassword() !== null) {
-            $currentUser->setPassword($updatedUser->getPassword());
-        }
         $errors = $validator->validate($currentUser);
         if (count($errors) > 0) {
             $jsonErrors = $serializer->serialize($errors, 'json');
@@ -171,28 +204,6 @@ class UserController extends AbstractController
         if ($customer !== null) {
             $cache->invalidateTags(['usersCache' . $currentUser->getCustomer()->getId()]);
         }
-
-        $content = $request->toArray();
-        $idCustomer = $content['customerId'] ?? null;
-        $password = $content['password'] ?? null;
-
-        //mot de passe modifié
-        if ($password !== null) {
-            $currentUser->setPassword(
-                $passwordHasher->hashPassword(
-                    $currentUser,
-                    $content['password']
-                )
-            );
-        }
-
-        //Customer modifié
-        if ($idCustomer !== null) {
-            $newCustomer = $customerRepository->find($idCustomer);
-            $currentUser->setCustomer($newCustomer);
-        }
-
-        $cache->invalidateTags(['usersCache' . $currentUser->getCustomer()->getId()]);
 
         $em->persist($currentUser);
         $em->flush();
@@ -219,52 +230,41 @@ class UserController extends AbstractController
      *     @OA\JsonContent(
      *          example={
      *              "email": "email@test.fr",
-     *              "password": "password",
-     *         "firstName": "firstName",
-     *     "lastName": "lastName",
-     *     "customerId": 1
+     *              "firstname": "firstname",
+     *              "lastname": "lastname",
      *          },
      * )
      * )
-     * @OA\Parameter(
-     *     name="email",
-     *     in="query",
-     *     required=true,
-     *     description="Email de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="password",
-     *     in="query",
-     *     required=true,
-     *     description="Mot de passe de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="firstName",
-     *     in="query",
-     *     description="Prénom de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="lastName",
-     *     in="query",
-     *     description="Nom de l'utilisateur",
-     *     @OA\Schema(type="string")
-     * )
-     * @OA\Parameter(
-     *     name="customerId",
-     *     in="query",
-     *     description="Id du customer",
-     *     @OA\Schema(type="integer")
-     * )
+     * @OA\Response(
+     *        response=401,
+     *        description="Non authentifié",
+     *    )
+     * @OA\Response(
+     *      response=403,
+     *      description="Accès refusé",
+     *      )
+     * @OA\Response(
+     *        response=404,
+     *        description="Resource non trouvée",
+     *    )
+     * @OA\Response(
+     *      response=500,
+     *      description="Erreur serveur",
+     *      )
      */
     #[Route('api/users', name: 'user_create', methods: ['POST'])]
-    public function createUser(Request $request, SerializerInterface $serializer, EntityManagerInterface $em, UserPasswordHasherInterface $passwordHasher, CustomerRepository $customerRepository, ValidatorInterface $validator, UrlGeneratorInterface $urlGenerator, TagAwareCacheInterface $cache): JsonResponse
-    {
+    public function createUser(
+        Request                $request,
+        SerializerInterface    $serializer,
+        EntityManagerInterface $em,
+        ValidatorInterface     $validator,
+        UrlGeneratorInterface  $urlGenerator,
+        TagAwareCacheInterface $cache
+    ): JsonResponse {
 
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
         $errors = $validator->validate($user);
+
         if (count($errors) > 0) {
             $jsonErrors = $serializer->serialize($errors, 'json');
             return new JsonResponse(
@@ -274,20 +274,17 @@ class UserController extends AbstractController
                 true
             );
         }
-        $user->setPassword(
-            $passwordHasher->hashPassword(
-                $user,
-                $user->getPassword()
-            )
-        );
-        $content = $request->toArray();
-        $idCustomer = $content['customerId'] ?? null;
-        $user->setCustomer($customerRepository->find($idCustomer));
+
+        $user->setCustomer($this->getUser());
         $user->setRoles(['ROLE_USER']);
         $em->persist($user);
         $em->flush();
         $cache->invalidateTags(['usersCache' . $user->getCustomer()->getId()]);
-        $location = $urlGenerator->generate('app_user_detail', ['id' => $user->getId()], UrlGeneratorInterface::ABSOLUTE_URL);
+        $location = $urlGenerator->generate(
+            'app_user_detail',
+            ['id' => $user->getId()],
+            UrlGeneratorInterface::ABSOLUTE_URL
+        );
         $context = SerializationContext::create()->setGroups(['getUserDetail']);
         $jsonUser = $serializer->serialize($user, 'json', $context);
 
@@ -308,27 +305,54 @@ class UserController extends AbstractController
      *     @OA\Items(ref=@Model(type=User::class, groups={"getUserDetail"})))
      *  )
      * )
-     *  @OA\Parameter(
+     * @OA\Parameter(
      *      name="page",
      *      in="query",
      *      description="Page que vous souhaitez récupérer",
      *      @OA\Schema(type="int")
      *  )
-     *  @OA\Parameter(
+     * @OA\Parameter(
      *      name="limit",
      *      in="query",
      *      description="Nombre d'éléments par page que vous souhaitez récupérer",
      *      @OA\Schema(type="int")
      *  )
+     * @OA\Response(
+     *         response=401,
+     *         description="Non authentifié",
+     *     )
+     * @OA\Response(
+     *       response=403,
+     *       description="Accès refusé",
+     *       )
+     * @OA\Response(
+     *         response=404,
+     *         description="Resource non trouvée",
+     *     )
+     * @OA\Response(
+     *     response=500,
+     *     description="Erreur serveur",
+     *     )
      *
      * @OA\Tag(name="Customers")
      */
     #[Route('api/customers/{id}/users', name: 'customer_users', methods: ['GET'])]
     public function getCustomerUsers(Customer $customer, UserRepository $userRepository, SerializerInterface $serializer, Request $request, TagAwareCacheInterface $cache): JsonResponse
     {
-        $page = $request->query->get('page', 1);
-        $limit = $request->query->get('limit', 3);
 
+        if ($customer !== $this->getUser()) {
+            return new JsonResponse(
+                null,
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
+        $page = $request->query->get('page', 1);
+
+        $limit = $request->query->get('limit', 3);
+        if ($limit > 10) {
+            $limit = 10;
+        }
         $idCache = 'customer-users-list-' . $customer->getId() . '-' . $page . '-' . $limit;
 
         $jsonUsers = $cache->get($idCache, function (ItemInterface $item) use ($userRepository, $customer, $page, $limit, $serializer) {
